requirements.txt
slack-sdk>=3.23.0
anthropic>=0.39.0
click>=8.1.7
python-dotenv>=1.0.0
rich>=13.7.0
flask>=3.0.0
requests>=2.31.0
